fixed "cap bonus > tower level" problem
The issue was, for example, you could initially have a 10 skyscraper bonus with a 10 cap and a 9 cap, but assume you reduced the skyscraper bonus to 6, the caps would stay the same, resulting in an inaccurate score. This just makes things a lot more logical, as if you reduce the skyscraper to 6, both caps would be brought down to 6.